refactor: MetricsPage.tsx の分析コンポーネントを metrics/ ディレクトリに分離
ToolDurationRanking, ContextUsageAnalysis, PromptWorkflowAnalysis,
EventsSection の4コンポーネントと共通のフォーマッター・定数を
frontend/src/components/metrics/ に抽出。
MetricsPage.tsx はオーケストレーターとして維持。

Closes #148
