+++ frontend/src/components/metrics/constants.ts
@@ -1,0 +1,15 @@
+export const EVENT_LABELS: Record<string, string> = {
+  "tool_result": "Tool Result",
+  "tool_decision": "Tool Decision",
+  "api_request": "API Request",
+  "api_error": "API Error",
+  "user_prompt": "User Prompt",
+};
+
+export const EVENT_COLORS: Record<string, string> = {
+  "tool_result": "bg-blue-100 text-blue-700",
+  "tool_decision": "bg-purple-100 text-purple-700",
+  "api_request": "bg-green-100 text-green-700",
+  "api_error": "bg-red-100 text-red-700",
+  "user_prompt": "bg-yellow-100 text-yellow-700",
+};
